requirements.txt
py-cord >= 2.1.3
asyncio >= 3.4.3
pyyaml >= 6.0
requests